requirements.txt
###### Requirements without Version Specifiers ######
paho-mqtt
readchar
FileCache
json